# newsboat: run parallel build for 'make' commands