@trofi 2023-01-08
Result of `nixpkgs-review pr 209659` run on x86_64-linux [1](https://github.com/Mic92/nixpkgs-review)
<details>
  <summary>1 package built:</summary>
  <ul>
    <li>newsboat</li>
  </ul>
</details>