@trofi 2023-01-09
AFAIU `enableParallelBuilding = true;` has no effect on manually called `make` calls. Setting it makes no difference for me.

From what I see when I watch interactive `newsboat` already builds and tests rust code in parallel.

- `cargo build` is a `cargo`'s default.
- `cargo check` is a `nixpkgs` default (unless opted out): https://github.com/NixOS/nixpkgs/blob/master/pkgs/build-support/rust/hooks/cargo-check-hook.sh#L13